enabled features implemented
